+++ src/main/java/lotto/controller/LottoController.java
@@ -2,15 +2,15 @@ package lotto.controller;
 
 import java.util.List;
 import java.util.function.Supplier;
-import lotto.Lotto;
-import lotto.LottoMachine;
-import lotto.LottoNumber;
-import lotto.LottoResult;
-import lotto.Lottos;
-import lotto.Money;
-import lotto.NumberPicker;
-import lotto.RandomNumberPicker;
-import lotto.WinningNumber;
+import lotto.model.Lotto;
+import lotto.model.LottoMachine;
+import lotto.model.LottoNumber;
+import lotto.model.LottoResult;
+import lotto.model.Lottos;
+import lotto.model.Money;
+import lotto.model.NumberPicker;
+import lotto.model.RandomNumberPicker;
+import lotto.model.WinningNumber;
 import lotto.dto.LottoResponse;
 import lotto.view.input.InputView;
 import lotto.view.output.OutputView;
