+++ src/main/java/lotto/dto/LottoResponse.java
@@ -1,7 +1,7 @@
 package lotto.dto;
 
 import java.util.List;
-import lotto.LottoNumber;
+import lotto.model.LottoNumber;
 
 public record LottoResponse(List<LottoNumber> numbers) {
 }
