+++ src/main/java/lotto/model/Lottos.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.model;
 
 import java.util.Collections;
 import java.util.List;
