+++ src/main/java/lotto/model/RandomNumberPicker.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.model;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.Collections;
