+++ src/test/java/lotto/LottoResultTest.java
@@ -5,6 +5,9 @@ import static org.assertj.core.api.Assertions.assertThat;
 import java.util.HashMap;
 import java.util.Map;
 import lotto.base.TestSupport;
+import lotto.model.LottoPrize;
+import lotto.model.LottoResult;
+import lotto.model.Money;
 import org.junit.jupiter.api.Test;
 
 class LottoResultTest extends TestSupport {
@@ -20,15 +23,15 @@ class LottoResultTest extends TestSupport {
         // when & then
         assertThat(result.getCountByPrize(LottoPrize.FIRST)).isEqualTo(1);
         assertThat(result.getCountByPrize(LottoPrize.FIFTH)).isEqualTo(2);
-        assertThat(result.getCountByPrize(LottoPrize.SECOND)).isZero()
+        assertThat(result.getCountByPrize(LottoPrize.SECOND)).isZero();
     }
 
     @Test
     void 총_당첨_금액을_계산할_수_있다() {
         // given
         Map<LottoPrize, Long> prizeCount = new HashMap<>();
-        prizeCount.put(LottoPrize.FIFTH, 2L);  // 5,000 * 2 = 10,000
-        prizeCount.put(LottoPrize.FOURTH, 1L);  // 50,000 * 1 = 50,000
+        prizeCount.put(LottoPrize.FIFTH, 2L);
+        prizeCount.put(LottoPrize.FOURTH, 1L);
         LottoResult result = new LottoResult(prizeCount);
 
         // when
@@ -42,7 +45,7 @@ class LottoResultTest extends TestSupport {
     void 수익률을_계산할_수_있다() {
         // given
         Map<LottoPrize, Long> prizeCount = new HashMap<>();
-        prizeCount.put(LottoPrize.FIFTH, 1L);  // 5,000원
+        prizeCount.put(LottoPrize.FIFTH, 1L);
         LottoResult result = new LottoResult(prizeCount);
         Money investment = new Money(8_000);
