+++ src/test/java/lotto/LottoTest.java
@@ -7,6 +7,8 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.util.List;
 import lotto.base.TestSupport;
+import lotto.model.Lotto;
+import lotto.model.NumberPicker;
 import org.junit.jupiter.api.Test;
 
 class LottoTest extends TestSupport {
